% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACBD.R
\docType{class}
\name{ACBD-class}
\alias{ACBD-class}
\alias{ACBD}
\title{Concrete interface for ACBD designs}
\description{
  Specialized Concrete Definition.
}
\section{Fields}{
  
  \describe{
    \item{\code{report}}{list or analytical summary.}
  }}


\examples{
  data(stg12025, package = "yabaf")
  b <- Breeder(stg12025)
  l <- ACBD(b)
  l$report$fit$sigma
}
\seealso{
  [Breeder() and SplitPlot()] for the pure virtual and concrete
  abstract template.
}
