% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{class}
\name{Breeder-class}
\alias{Breeder-class}
\alias{Breeder}
\title{Virtual interface for mapped tabular objects}
\description{
Pure abstract Definition.
}
\section{Fields}{

\describe{
\item{\code{data}}{Tabulated data instance.}

\item{\code{log}}{Character vector history.}
}}


\examples{
data(std1997, package = "yabaf")
b <- Breeder(std1997)
class(b)
}
\seealso{
[SplitPlot()] for a the in-memory concrete implementation.
}
