% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{gg1984}
\alias{gg1984}
\title{Lattice Example Data Set}
\format{
data frame with four columns: replication, block, treat and response
}
\usage{
data(gg1984)
}
\description{
Square Lattice design example data set. Reproducible results.
    Check data and outputs in reference (\insertCite{gg;textual}{yabaf}).
}
\references{
\insertAllCited{}
}
