% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pedigree.R
\docType{data}
\name{ks1988}
\alias{ks1988}
\title{Pedigree Example Data Set}
\format{
data frame with four columns: replication, block, treat and response
}
\usage{
data(ks1988)
}
\description{
Pedigree example data set. Reproducible results.
    Check data and outputs in reference (\insertCite{ks;textual}{yabaf}).
}
\references{
\insertAllCited{}
}
