% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice.R
\docType{class}
\name{Lattice-class}
\alias{Lattice-class}
\alias{Lattice}
\title{Concrete interface for Lattice designs}
\description{
Specialized Concrete Definition.
}
\section{Fields}{

\describe{
\item{\code{report}}{list or analytical summary.}
}}


\examples{
data(gg1984, package = "yabaf")
b <- Breeder(gg1984)
l <- Lattice(b)
l$report$fit$sigma
}
\seealso{
[Breeder() and SplitPlot()] for the pure virtual and concrete
    abstract template.
}
