% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pedigree.R
\docType{class}
\name{Pedigree-class}
\alias{Pedigree-class}
\alias{Pedigree}
\title{Concrete interface for Pedigrees}
\description{
Specialized Concrete Definition.
}
\section{Fields}{

\describe{
\item{\code{structure}}{list with the additive relationship matrix.}
}}


\examples{
data(ks1988, package = "yabaf")
b <- Breeder(ks1988)
p <- Pedigree(b)
p$structure$additive
}
\seealso{
[Breeder() and SplitPlot()] for the pure virtual and concrete
    abstract template.
}
