% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCBD.R
\docType{class}
\name{RCBD-class}
\alias{RCBD-class}
\alias{RCBD}
\title{Concrete interface for Randomized Complete Blocks}
\description{
Specialized Concrete Definition.
}
\section{Fields}{

\describe{
\item{\code{report}}{list or analytical summary.}
}}


\examples{
data(ergoStool, package = "nlme")
names(ergoStool) <- c("response", "treat", "block")
b <- Breeder(as.data.frame(ergoStool))
r <- RCBD(b)
r$report$fit$sigma
}
\seealso{
[Breeder() and SplitPlot()] for the pure virtual and concrete
    abstract template.
}
