% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{class}
\name{SplitPlot-class}
\alias{SplitPlot-class}
\alias{SplitPlot}
\title{Concrete interface for Split Plot designs}
\description{
Sample of Concrete Definition.
}
\section{Fields}{

\describe{
\item{\code{report}}{list or analytical summary.}
}}


\examples{
data(std1997, package = "yabaf")
b <- Breeder(std1997)
s <- SplitPlot(b)
s$report$fit$sigma
}
\seealso{
[Breeder()] for the pure virtual abstract template.
}
