% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{std1997}
\alias{std1997}
\title{Example Data Set}
\format{
data frame with four columns: response, plot, sub and block
}
\usage{
data(std1997)
}
\description{
Split plot design example data set. Reproducible results.
    Check data and outputs in reference (\insertCite{std;textual}{yabaf}).
}
\references{
\insertAllCited{}
}
