% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{stg12025}
\alias{stg12025}
\title{ACBD Example Data Set}
\format{
  data frame with five columns: row, column, block, treat and response
}
\usage{
  data(stg12025)
}
\description{
  Augmented complete block design example data set. Reproducible results.
  Check data and outputs in reference (\insertCite{std;textual}{yabaf}).
}
\references{
  \insertAllCited{}
}
