% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trial.R
\docType{class}
\name{Trial-class}
\alias{Trial-class}
\alias{Trial}
\title{Concrete interface for Randomized Complete Blocks}
\description{
Specialized Concrete Definition.
}
\section{Fields}{

\describe{
\item{\code{report}}{list or analytical summary.}
}}


\examples{
data(gg1984, package = "yabaf")
gg1984$design <- "Lattice"
b1 <- Breeder(gg1984)
l <- Trial(b1)
l$report$fit$sigma
data(wheat, package = "yabaf")
wheat$design <- "Augmented"
b2 <- Breeder(wheat)
a <- ACBD(b2)
a$report$fit$sigma
}
\seealso{
[Breeder() and SplitPlot()] for the pure virtual and concrete
    abstract template.
}
