% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACBD.R
\docType{data}
\name{wheat}
\alias{wheat}
\title{ACBD Example Data Set}
\format{
data frame with six columns: row, column, block, treat, role and response
}
\usage{
data(wheat)
}
\description{
Augmented complete block design example data set. Reproducible results.
    Check data and outputs.
}
