% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{yabaf-package}
\alias{yabaf-package}
\alias{yabaf}
\title{yabaf: Yet Another Breeding Analytical Framework}
\description{
First phrase. Then the second. The quick brown fox jumps over
    the lazy dog.
}
\author{
Fernando H. Toledo \email{f.toledo@cgiar.org}
}
\keyword{package}
